requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.7
openai-whisper==20231117
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
torch==2.1.1
transformers==4.36.0
librosa==0.10.1
soundfile==0.12.1
ffmpeg-python==0.2.0
numpy>=1.24.0
scipy>=1.10.0
